refactor(formats): split alignment_io into package
